Rework comment re allowing comments and ci specific trigger
A comment seems always to be of the form
```
Patch Set \d+:

the actual comment
```

So, instead of matching the "suffix" of the comment, match on the
prefix of the string including the header, allowing for the prefix
`sap-openstack-ci` to `recheck` (i.e. `sap-openstack-ci recheck`) and
also allow arbitrary text after `recheck` for comments as long as it is
separated from the `recheck` by a whitespace.